blog-posts/agent-tech-risk/packages/risk-discovery/pyproject.toml
[project]
name = "risk-discovery"
version = "0.1.0"
description = "AWS Risk Discovery Agent for PE Technical Due Diligence"
requires-python = ">=3.12"
dependencies = [
    "risk-generator",
    "pydantic>=2.0",
    "pydantic-ai>=0.2",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
    "boto3>=1.35",
    "python-dotenv>=1.0",
    "huggingface-hub>=0.20",
]

[project.scripts]
risk-discovery = "risk_discovery.cli:main"

[tool.uv.sources]
risk-generator = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [".env", ".env.*", "*.pyc", "__pycache__"]

[tool.hatch.build.targets.wheel]
packages = ["src/risk_discovery"]
